# opt: Allow reuse of optimizer