Allow the optimizer data structures to be reused for optimizing
multiple queries. Also, use a different data structure for detecting
rule cycles in order to save creation of a map.

before:
```
name                      time/op
Phases/kv-read/OptBuild   18.8µs ± 2%
Phases/kv-read/Normalize  18.8µs ± 1%
Phases/kv-read/Explore    24.5µs ± 3%
```
after:
```
Phases/kv-read/OptBuild   18.0µs ± 1%
Phases/kv-read/Normalize  17.8µs ± 3%
Phases/kv-read/Explore    23.1µs ± 2%
```